+++ eo-maven-plugin/src/main/java/org/eolang/maven/dependencies/DcsWithRuntime.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
+package org.eolang.maven.dependencies;
 
 import com.jcabi.xml.XMLDocument;
 import java.io.IOException;
@@ -37,7 +37,7 @@ import org.cactoos.scalar.Unchecked;
  *
  * @since 0.28.11
  */
-final class DcsWithRuntime implements Iterable<Dependency> {
+public final class DcsWithRuntime implements Iterable<Dependency> {
 
     /**
      * Dependency downloaded by HTTP from Maven Central.
@@ -59,7 +59,7 @@ final class DcsWithRuntime implements Iterable<Dependency> {
      *
      * @param dlg Dependencies delegate.
      */
-    DcsWithRuntime(final Iterable<Dependency> dlg) {
+    public DcsWithRuntime(final Iterable<Dependency> dlg) {
         this(dlg, DcsWithRuntime.MAVEN_DEPENDENCY);
     }
 
@@ -69,7 +69,7 @@ final class DcsWithRuntime implements Iterable<Dependency> {
      * @param dlg Dependencies delegate.
      * @param sup Dependency.
      */
-    DcsWithRuntime(
+    public DcsWithRuntime(
         final Iterable<Dependency> dlg,
         final Dependency sup
     ) {
