+++ eo-maven-plugin/src/test/java/org/eolang/maven/dependencies/DcsJsonTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
+package org.eolang.maven.dependencies;
 
 import java.io.IOException;
 import java.nio.file.Path;
